requirements3.txt
flask==3.1.0
flask-swagger-ui==4.11.1
pymongo==4.10.1
python-dotenv==1.0.1
nltk==3.9.1
spacy==3.8.3
werkzeug==3.1.3
nltk==3.9.1
spacy==3.8.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
sumy==0.11.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.4
pytest-flask==1.3.0
flake8==7.1.1
black==24.10.0